blau fehlt

+++ framework/color.hpp
@@ -3,13 +3,13 @@
 
 struct Color
 {
-	double r;
-	double g;
-	double b;
+	float r;
+	float g;
+	float b;
 
-	Color(double x, double y, double z) : r{x}, g{y}, b{z} {}
+	Color(float x, float y, float z) : r{x}, g{y}, b{z} {}
 
-	Color(double x) : r{x}, g{x}, b{x} {}
+	Color(float x) : r{x}, g{x}, b{x} {}
 
 	Color() :  r{0.0}, g{0.0}, b{0.0} {}
 };
